delegate page-related logic from BoardList to BoardsPage
